UPSTREAM: drm/amd/display: Correct unit conversion for vstartup
[why]
vstartup is calculated to be a large number. it works because
it is within vertical blank, but it reduces region of blank that
can be used for power gating.

[how]
Calculation needs to convert micro seconds to number of
vertical lines.

Reviewed-by: Kazlauskas Nicholas <Nicholas.Kazlauskas@amd.com>
Acked-by: Wayne Lin <wayne.lin@amd.com>
Signed-off-by: Reza Amini <reza.amini@amd.com>
Tested-by: Daniel Wheeler <daniel.wheeler@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit 8f31c7be3e89634767686ffab869cd3ebfea188f)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=323757777
TEST=cq dry run

Backported using forklift.py

Change-Id: I4fa868ce1a31aec47d19a529f0fe8d2274efa290