UPSTREAM: drm/i915/huc: check HuC and GuC version compatibility on MTL
Due to a change in the auth flow on MTL, GuC 70.7.0 and newer will only
be able to authenticate HuC 8.5.1 and newer. The plan is to update the 2
binaries synchronously in linux-firmware so that the fw repo always has
a matching pair that works; still, it's better to check in the kernel so
we can print an error message and abort HuC loading if the binaries are
out of sync instead of failing the authentication.

v2: Add clarification comment, fix typo in commit msg, clean up variable
declaration (John)

Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Cc: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: Andrzej Hajda <andrzej.hajda@intel.com> #v1
Reviewed-by: John Harrison <John.C.Harrison@Intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20230717234905.117114-1-daniele.ceraolospurio@intel.com
(cherry picked from commit e4731b51c847e09ff87d50d6e4cd7ed7501978e9)
Signed-off-by: Lee Shawn C <shawn.c.lee@intel.com>

BUG=b:293823364
TEST= No intel_gt_init error while running warm reboot stress.

Change-Id: I6ad3a47393fda6d9dad635df122734c9252c6123
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4864366
Tested-by: Shawn C Lee <shawn.c.lee@intel.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Shawn C Lee <shawn.c.lee@intel.com>