UPSTREAM: drm/i915: Start using plane scale factor for relative data rate
BSpec clearly instructs us to use plane scale factor when calculating
relative data rate to be used when allocating DDB blocks for each plane.
For some reason we use scale factor for data_rate calculation, which is
used for BW calculations, however we are not using it for DDB calculations.
So lets fix it as described in BSpec 68907.

Signed-off-by: Stanislav Lisovskiy <stanislav.lisovskiy@intel.com>
Reviewed-by: Garg, Nemesa <nemesa.garg@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20230719104833.25366-1-stanislav.lisovskiy@intel.com
(cherry picked from commit a86c75dcdd0374444514c1e40411177ff7afe9bd)

BUG=b:279496283
TEST=tast.graphics.IgtKms.kms_cursor_legacy test should get pass

Change-Id: I407a3174b3a3859a6064e1b21a9af98c95a9d663
Signed-off-by: Rakshith M O <rakshith.m.o@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5224447
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Waheem Md <waheem.md@intel.corp-partner.google.com>