CHROMIUM: iwl7000: mvm: don't add dummy phy context
From its very first stages of development, iwlmvm added all the PHY
context immediately upon firmware boot. Then, all we needed to do is to
modify the contexts. This was fine if the addition of a PHY context that
we don't need is free. This was true until now. Newer devices will run
calibrations upon the addition of a PHY context.

Change the way we work with PHY context in iwlmvm. Fortunately, we
already have all the ref counting in place so that it is not very hard
to do.

Also, since we now remove the PHY context before the link is removed
(but after it has been de-activated of course), it'll confuse the
firmware if we put the late phy_id into the LINK command that removes
the link. Change this to put an invalid phy_id just like we do when we
add a link that has no PHY context yet.

Change-Id: I2032a7d227b57f4fc4370a2793476d47538404fd
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
tested: iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: 4127db9168c6f2e160d666097f012640c9618066