UPSTREAM: drm/i915/guc: Force a reset on internal GuC error
If GuC hits an internal error (and survives long enough to report it
to the KMD), it is basically toast and will stop until a GT reset and
subsequent GuC reload is performed. Previously, the KMD just printed
an error message and then waited for the heartbeat to eventually kick
in and trigger a reset (assuming the heartbeat had not been disabled).
Instead, force the reset immediately to guarantee that it happens and
to eliminate the very long heartbeat delay. The captured error state
is also more likely to be useful if captured at the time of the error
rather than many seconds later.

Note that it is not possible to trigger a reset from with the G2H
handler itself. The reset prepare process involves flushing
outstanding G2H contents. So a deadlock could result. Instead, the G2H
handler queues a worker thread to do the reset asynchronously.

v2: Flush the worker on suspend and shutdown. Add rate limiting to
prevent spam from a totally dead system (review feedback from Daniele).

Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20230816003957.3572654-1-John.C.Harrison@Intel.com
(cherry picked from commit b2edc4148ad90c906275329711810721719a0c2d)
Signed-off-by: Lee Shawn C <shawn.c.lee@intel.com>

BUG=b:293823364
TEST= No intel_gt_init error while running warm reboot stress.

Change-Id: Ic037f30fd8d75f93bd78f07f8583bb972a1eee78
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5104866
Commit-Queue: Shawn C Lee <shawn.c.lee@intel.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Tested-by: Shawn C Lee <shawn.c.lee@intel.com>