CHROMIUM: iwl7000: mvm: Return success if link could not be removed
iwl_mvm_remove_link would return an error if the link could not be
removed. However, doing so prevents recovery if a link was not uploaded
to the FW in the first place and the link_info was not allocated or
fw_link_id is not set.

Returning success means that we can still try to continue with adding
new links in change_vif_links.

Change-Id: Idc8fbd671362d962c02b1df87fa6258733631580
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
tested: iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: b319d628e9b8681f0d8e28fb702f981304920452