UPSTREAM: drm/i915/display: Eliminate IS_METEORLAKE checks
Most of the IS_METEORLAKE checks in the display code shouldn't actually
be tied to MTL as a platform, but rather to the Xe_LPD+ display IP
(which is used in MTL, but may show up again in future platforms).  In
cases where we're trying to match that specific IP, use a version check
against IP_VER(14, 0).  For cases where we're just handling new behavior
introduced by this IP (but which may also be inherited by future IP as
well), use a ver >= 14 check.

The one exception here is the stolen memory workaround Wa_13010847436
(which is mislabelled as "Wa_22018444074" in the code).  That's truly a
MTL-specific issue rather than being tied to any of the IP blocks, so
leaving the condition as IS_METEORLAKE is correct there.

v2:
 - cdclk check should be >=, not >.  (Gustavo)

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: Gustavo Sousa <gustavo.sousa@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20230821180619.650007-19-matthew.d.roper@intel.com
(cherry picked from commit e388ae97e2253363275396b0f74c0df7d3f03b15)

BUG=b:314887266
TEST=i915 works, no driver errors, no visible issues or flickers

Change-Id: I41fd96de176494acf2ace08e1624e5eeb7cf3187
Signed-off-by: Paz Zcharya <pazz@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5244368
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Paz Zcharya <pazz@google.com>
Tested-by: Paz Zcharya <pazz@google.com>
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>