UPSTREAM: ASoC: SOF: ipc4-pcm: fixup dailink based on copier format
When a copier exposes a single format, we can fixup the BE dailink with
that format. This is helpful when some codec have format restrictions and
e.g. don't support a 32-bit format. In that case, the copier output
formats mirror that restriction in the topology file.

An alternate solution was suggested earlier using a dedicated topology
token. When specified, the token would be used to fix-up the dailink. The
main reason why this solution was chosen is that there is a risk of a
disconnect between token definition and copier format. With a single piece
of information as suggested in this patch, there are fewer risks of a bad
configuration.

Conflicts:
    sound/soc/sof/ipc4-pcm.c

Due to 'ASoC: sof: convert not to use asoc_xxx()' unified function conversion.

Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>
Reviewed-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
Link: https://lore.kernel.org/r/20230915093507.7242-3-peter.ujfalusi@linux.intel.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 26dfc43461102957e33454e766d592df330ef7a0
 https://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git for-next)

BUG=b:327575760
TEST=test all audio devices are functional.

Change-Id: Ib67be4760588823ae5114defb51d5be84b073e07
Signed-off-by: Mac Chiang <mac.chiang@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5317290
Commit-Queue: Pin-chih Lin <johnylin@google.com>
Reviewed-by: Pin-chih Lin <johnylin@google.com>