BACKPORT: wifi: cfg80211: report per-link errors during association
When one of the links (other than the assoc_link) is misconfigured
and cannot work the association will fail. However, userspace was not
able to tell that the operation only failed because of a problem with
one of the links. Fix this, by allowing the driver to set a per-link
error code and reporting the (first) offending link by setting the
bad_attr accordingly.

This only allows us to report the first error, but that is sufficient
for userspace to e.g. remove the offending link and retry.

Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
Signed-off-by: Gregory Greenman <gregory.greenman@intel.com>
Link: https://lore.kernel.org/r/20230920211508.ebe63c0bd513.I40799998f02bf987acee1501a2522dc98bb6eb5a@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit a7b2cc591d556eca044c823e4e92d8c1cb430ef7)
Signed-off-by: Matthew Wang <matthewmwang@chromium.org>

Conflicts:
	net/wireless/nl80211.c - about 40 changes were made to move the
	locking mechanism. The backport would have required quite a bit
	of reverting and backporting, so I left it as is.

BUG=b:327409698
TEST=matfunc on karis

Change-Id: I940876433c23fc39a659d5aa1d9e1fc57ea85ded
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5335601
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Johannes Berg <johannes.berg@intel.corp-partner.google.com>
Reviewed-by: David Ruth <druth@chromium.org>