FROMGIT: drm/i915: Lift runtime-pm acquire callbacks out of intel_wakeref.mutex
When runtime pm is first woken, it will synchronously call the
registered callbacks for the device. These callbacks
may pull in their own forest of locks, which we do not want to
conflate with the intel_wakeref.mutex. A second minor benefit to
reducing the coverage of the mutex, is that it will reduce
contention for frequent sleeps and wakes (such as when being used
for soft-rc6).

v2: remove usage of fetch_and_zero() and other improvements(Jani)

Signed-off-by: Chris Wilson <chris.p.wilson@intel.com>
Signed-off-by: Nirmoy Das <nirmoy.das@intel.com>
Reviewed-by: Andi Shyti <andi.shyti@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20230926083742.14740-2-nirmoy.das@intel.com
(cherry picked from commit 4cd64e9d2c7206db05e7162d0258b455726b7ec5
 https://anongit.freedesktop.org/git/drm-intel drm-intel-gt-next)

BUG=b:298527509
TEST=Run S0ix test for 2500 cycles, No kernel panic should be observed

Change-Id: Ie7754c569c7b28de8d2e483706c27b58ddd3667e
Signed-off-by: Rakshith M O <rakshith.m.o@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4955118
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Tested-by: Arselan Alvi <arselan.alvi@intel.com>