CHROMIUM: sched/fair: Make minimum load balance interval a tunable
We are seeing a several fold reduction in number of SCHED_SOFTIRQ by
tuning the minimum interval of periodic load balance to a higher value.

This could also result in higher power savings, lower CPU usage and
lesser system overhead. Make it a tunable so that it is easier for
userspace to experiment, such as via finch on ChromeOS.

While this effects both high and low resolution timers, this will bring
down the recent increase in softirqs seen in low resolution timers as
well. There are other patches also that fix the increase in softirq
specifically for low resolution timers.

BUG=b:263289152
UPSTREAM-TASK=b:303083202

Co-developed-by: Vineeth Pillai <vineethrp@google.com>
Change-Id: I5041e0a01bd11e617d59d8f8d23c2f59d5772f47
Signed-off-by: Joel Fernandes <joelaf@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4908115
Reviewed-by: Vineeth Pillai <vineethrp@google.com>