FROMGIT: drm/dp_mst: Swap the order of checking root vs. non-root port BW limitations
drm_dp_mst_atomic_check_mgr() should check for BW limitation starting
from sink ports continuing towards the root port, so that drivers can
use the @failing_port returned to resolve a BW overallocation in an
ideal way. For instance from streams A,B,C in a topology A,B going
through @failing_port and C not going through it, a BW overallocation of
A,B due to a limit of the port must be resolved first before considering
the limits of other ports closer to the root port. This way can avoid
reducing the BW of stream C unnecessarily due to a BW limit closer to the
root port.

Based on the above swap the order of the BW check for the root port and
the check for all the ports downstream of it (the latter going through
the topology already in the sink->root port direction).

Cc: Lyude Paul <lyude@redhat.com>
Cc: dri-devel@lists.freedesktop.org
Reviewed-by: Lyude Paul <lyude@redhat.com>
Reviewed-by: Stanislav Lisovskiy <stanislav.lisovskiy@intel.com>
Acked-by: Maxime Ripard <mripard@kernel.org>
Signed-off-by: Imre Deak <imre.deak@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20231030155843.2251023-6-imre.deak@intel.com
(cherry picked from commit d075bca47c18779301fee5a9d140f146cde4b532
 https://anongit.freedesktop.org/git/drm-intel.git drm-intel-next)

BUG=b:298062999
TEST=DP2.1 displays should work fine

Change-Id: I92b57967c20f6435585aad8ea707d88ac7bb65e0
Signed-off-by: Ap, Kamal <kamal.ap@intel.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5054907
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Tested-by: Natarajan Balamurugan <balamurugan.natarajan@intel.corp-partner.google.com>
Reviewed-by: Sean Paul <sean@poorly.run>