FROMGIT: drm/i915/dp_mst: Add atomic state for all streams on pre-tgl platforms
If an MST stream is modeset, its state must be checked along all the
other streams on the same MST link, for instance to resolve a BW
overallocation of a non-sink MST port or to make sure that the FEC is
enabled/disabled the same way for all these streams.

To prepare for that this patch adds all the stream CRTCs to the atomic
state and marks them for modeset similarly to tgl+ platforms. (If the
state computation doesn't change the state the CRTC is switched back to
fastset mode.)

So far on tgl+ this was required because all streams in the topology
shared the master transcoder. For older platforms this didn't apply but
adding all the state is required now on all platforms based on the
above.

v2:
- Add code and commit log comment clarifying the requirements on old/new
  platforms. (Stan)
- Rename the function based on the new semantic. (Ville)

Cc: Ville Syrjälä <ville.syrjala@linux.intel.com>
Reviewed-by: Stanislav Lisovskiy <stanislav.lisovskiy@intel.com>
Signed-off-by: Imre Deak <imre.deak@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20231030155843.2251023-15-imre.deak@intel.com
(cherry picked from commit 5d78cd80efdd4ac221a0ccd884082280ddef6128
 https://anongit.freedesktop.org/git/drm-intel.git drm-intel-next)

BUG=b:298062999
TEST=DP2.1 displays should work fine

Change-Id: I3a88e552bce0d725219168ed16e7249401ccb94e
Signed-off-by: Ap, Kamal <kamal.ap@intel.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5054916
Tested-by: Natarajan Balamurugan <balamurugan.natarajan@intel.corp-partner.google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Drew Davenport <ddavenport@chromium.org>