FROMGIT: drm/i915/dp_mst: Improve BW sharing between MST streams
At the moment modesetting a stream CRTC will fail if the stream's BW
along with the current BW of all the other streams on the same MST link
is above the total BW of the MST link. Make the BW sharing more dynamic
by trying to reduce the link bpp of one or more streams on the MST link
in this case.

When selecting a stream to reduce the BW for, take into account which
link segment in the MST topology ran out of BW and which streams go
through this link segment. For instance with A,B,C streams in the same
MST topology A and B may share the BW of a link segment downstream of a
branch device, stream C not downstream of the branch device, hence not
affecting this BW. If this link segment's BW runs out one or both of
stream A/B's BW will be reduced until their total BW is within limits.

While reducing the link bpp for a given stream DSC may need to be
enabled for it, which requires FEC on the whole MST link. Check for this
condition and recompute the state for all streams taking the FEC
overhead into account (on 8b/10b links).

v2:
- Rebase on s/min_bpp_pipes/min_bpp_reached_pipes/ change.

Reviewed-by: Stanislav Lisovskiy <stanislav.lisovskiy@intel.com>
Signed-off-by: Imre Deak <imre.deak@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20231030155843.2251023-29-imre.deak@intel.com
(cherry picked from commit 36f579ffc6921408fd2e466a6930463bac56b926
 https://anongit.freedesktop.org/git/drm-intel.git drm-intel-next)

BUG=b:298062999
TEST=DP2.1 displays should work fine

Signed-off-by: Ap, Kamal <kamal.ap@intel.corp-partner.google.com>
Change-Id: I1cca7f9ebccde9f84c29693bec81c29338d8ad5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5054930
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Tested-by: Natarajan Balamurugan <balamurugan.natarajan@intel.corp-partner.google.com>